* test/win32ole/test_err_in_callback.rb (teardown): remove tmp file   only when it exists.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@29924 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ test/win32ole/test_err_in_callback.rb
@@ -7,7 +7,6 @@ begin
   require 'win32ole'
 rescue LoadError
 end
-require 'rbconfig'
 if defined?(WIN32OLE)
   require 'mkmf'
   require 'test/unit'
@@ -20,7 +19,7 @@ if defined?(WIN32OLE)
         @iopt = $:.map {|e|
           " -I " + e
         }.join("")
-        @script = File.dirname(__FILE__) + "/err_in_callback.rb"
+        @script = File.join(File.dirname(__FILE__), "err_in_callback.rb")
       end
     end
 
@@ -37,8 +36,7 @@ if defined?(WIN32OLE)
     end
 
     def teardown
-      File.unlink("test_err_in_callback.log")
+      File.unlink("test_err_in_callback.log") if File.exist?("test_err_in_callback.log")
     end
-
   end
 end
